@@ -20,13 +20,13 @@
   "prefer-stable": true,
   "config": {
     "platform": {
-      "php": "5.6.4"
+      "php": "7.1.3"
     },
     "prepend-autoloader": false
   },
   "require": {
-    "php": ">=5.6.4",
-    "jtl/connector": "^2.4.0",
+    "php": ">=7.1.3",
+    "jtl/connector": "^3.2.0",
     "symfony/yaml": "^3.4",
     "ext-json": "*",
     "phing/phing": "^2.16"
