@@ -15,10 +15,11 @@ class ImageI18n extends BaseController
         $i18ns = [];
 
         if ($data['relationType'] === 'product') {
-
             $imageLang = $this->db->executeS(
-                sprintf('SELECT i.id_lang, i.legend as altText FROM ' . _DB_PREFIX_ . 'image_lang i WHERE i.id_image = %s',
-                    $data['id'])
+                sprintf(
+                    'SELECT i.id_lang, i.legend as altText FROM ' . _DB_PREFIX_ . 'image_lang i WHERE i.id_image = %s',
+                    $data['id']
+                )
             );
 
             if (is_array($imageLang)) {
