@@ -361,7 +361,7 @@ class Product extends BaseController
     private function pullSpecialAttributes($data, $model)
     {
         foreach (ProductAttr::getSpecialAttributes() as $wawiName => $prestaName) {
-            if(isset($data[$prestaName])) {
+            if (isset($data[$prestaName])) {
                 $attribute = new \jtl\Connector\Model\ProductAttr();
                 $attributeI18n = new \jtl\Connector\Model\ProductAttrI18n();
                 $attribute->setId(new Identity($prestaName));
@@ -371,11 +371,11 @@ class Product extends BaseController
                 $attributeI18n->setName($wawiName);
 
                 $value = $data[$prestaName];
-                if($wawiName === 'main_category_id') {
+                if ($wawiName === 'main_category_id') {
                     $value = (string)$this->findCategoryHostIdByEndpoint($data[$prestaName]);
                 }
 
-                if($value !== '') {
+                if ($value !== '') {
                     $attributeI18n->setValue($value);
                     $attribute->setI18ns([$attributeI18n]);
                     $model->addAttribute($attribute);
